[FEATURE] Add quicklinks on startpage
The current startpage was copied from some other manual. Until recently the meta
information on it did not exist and no one seemed to miss it.

This change makes some extra room on start page by removing information that is
redundant (is in footer or elsewhere anyway) and provides quick links as already
is common practice on main docs page: https://docs.typo3.org/#common-quick-links.

Resolves: #56