Add information about pushing to refs/for/master again
We add the full push command again for pushing to master:
git push origin HEAD:refs/for/master

This was simplified in previous PR as the full command is
not necessary if already configured.

However, this change was made too abrupt and does not give
people the time to update their configuration and adjust
their workflow.

So we added the (previous) command again with a hint and a
link to the configuration section.

Pushing to master was done like this:

  git push origin HEAD:refs/for/master

This can be simplified by:

  git push

if already configured like this:
  git config remote.origin.push +refs/heads/master:refs/for/master

Related: #167