Update Setup > Prerequisites
To streamline the contribution guide, some sections for installing
and executing the tools are removed or changed.

Specifically, the listed tool installation on the prerequisites
page can be found elsewhere (and will usually be more up-to-date
elsewhere).

Explanation:

Previously, the setup and tool installation existed in various
places, sometimes with differening instructions. For example,
the Prerequisites page included instructions for installing
yarn on Mac OSX, but there was also a page in the appendix with
different instructions. Additionally, the installation instructions
for the platforms are  easy to find elsewhere and are more up-to-date
on the documentation pages of these tools.

On top of this, dependency on tool installation is no longer a
prerequisite if runTests.sh is used (or even if DDEV is used without
runTests.sh).

Related: #253
Related: #249