Reflect other responsibilities of runTests.sh
This change converts the page "Run tests" to be about runTests.sh
in general - including running the tests but not restricted to this.

The page "Run tests" used to be just about running test. However,
more and more commands for checking, fixing, building etc. have
since been added to the script runTests.sh.