Remove video from DDEV page (#251)
The video explains setting up a TYPO3 site with DDEV, but
not for core development. This means, it is slightly different,
e.g. TYPO3 is intalled via Composer and not based on a cloned
git repository of TYPO3.

In the docs, currently, it is recommended to watch the video
OR follow the steps below. This is misleading, as following
the video will not get you a TYPO3 site ready to be used for
core development.

While the video provides some more in-depth information, this
is not that much relevant for setting up TYPO3 for core development.

Most likely, in the future, there will be further changes for setting
up DDEV (e.g. use runTest.sh) and the page will continuously be
updated, but the video will not.