Remove setup steps from "Create a patch" (#266)
The "Create a patch" page was a complete walkthrough of all steps.
This had the advantage that you had all steps on one page, but
it basically repeated the setup steps which are already described
in the setup chapter.

The main reason for this, was that the setup chapter was quite
lengthy. But in this case, it would be better to have the lengthy
part in setup streamlined and possibly some parts removed or moved
to appendix.

Repeating the commands led to commands described in several places
which makes it difficult to maintain.