Add commands with runTests.sh for building (#265)
* Add commands with runTests.sh for building

Add the commands for building with runTests.sh:

- runTests.sh -s composerInstall
- runTests.sh -s buildCss
- runTests.sh -s buildJavascript

The section on yarn in the setup chapter was removed. It is not
necessary to build the assets when setting up. It is only necessary
when changing SCSS or Typescript / JavaScript etc. Additionally,
it is planned to have a section with all build commands in the
Howto section. This makes it more appropriate to remove the part
in the setup section. (Building is something we need later too,
not just in setup).

For all listed commands, we add TABs so that we can always show
the runTests.sh commands and the "direct commands" side by side.

This will hopefully make it less confusing for the readers.

It was also discussed to remove the direct commands entirely - however
there is sometimes still the need to use the direct command. In any
case it is always noted that runTests.sh is recommended.

Resolves: #249
Resolves: #254

* Update Documentation/Appendix/SettingUpTypo3Ddev.rst

Co-authored-by: Jonas Eberle <flightvision@googlemail.com>

* Update Documentation/Appendix/SettingUpTypo3Ddev.rst

Co-authored-by: Jonas Eberle <flightvision@googlemail.com>

Co-authored-by: Jonas Eberle <flightvision@googlemail.com>