Do not use specific PHP version in installation
The installation instructions are generalized.

Here, we do not specify a specific version to install as this
must constantly be updated and changed. The installation
instruction should be very minimal and refer back to the
system requirements page.

Additionally, add a note that adding a PPA may be necessary.

We refer to the DDEV page, in case someone is having problems with
the installation, they can use that as a more easy way to setup.

Resolves: #273