Fix code block style
Use recommended long form for code block style,
e.g. .. code-block:: bash instead of short form ::.

As documented, this is clearer.

Additionally, separate the command and result and use captions.

Remove the $ at the beginning of some commands as that was done
inconsistently and is usually not done elsewhere in this guide.

Resolves: #278