Move note about migtrating Git setup
The note is moved from the Git setup to the cheatsheet and split
into 2 parts with individual sections.

This moves the information from the Git setup which is for new
people (but there is still a much shorter note with the links
in case someone is looking for this).

The problem was that the note was very look, and it was not
very clear to contributors that they do not need to execute
this.

Resolves: #287