Clarify current directory for installing TYPO3 (#300)
There was a reference to htdocs directory which was not correct.

A core setup installation is described. This does not have a
htdocs or public directory. The FIRST_INSTALL must be
created in the main directory.

We use the directory /var/www/t3coredev which was introduced
in the description of the Git setup.