[TASK] Move the information on Core testing
It belongs more logically into contribution guide then in TYPO3 explained as TYPO3 explained aims at
the consuming developers, not the core developers.

Also we should have all information on Core testing in one place and consolidate it with time.

For now I moved the files as they were, just fixing references and indentation.

We should do changes in a follow up.