[TASK] Prevent using "no/low brainer" as wording
To quote a post on the typo3-cms-coredev slack on the
intentions of this, which I fully support:

There is nothing like a “no-brainer” or "low-brainer".
Every patch needs to be checked thoroughly.
The word “no-brainer” leads to hasty mergers.
There is a reason why something has to be changed or adapted.
If a change is so obvious, it would have been solved correctly from the very beginning.
There is always a chance that the “no-brainer” is broken anyway.
Please let us use our brains, always. Thank you.