Remove page "Resources for Editors" (#378)
This page was initially introduced to provide some information and
conventions for people editing this guide. However, the page is
only valualbe, if visible and commonly used.

Unless a general convention is established if and how to provide
information about editing a specific guide, the page is of little
value and should be removed.

Related: #255